+++ R/boot2est_lavaan.R
@@ -214,7 +214,11 @@ fit2boot_out_do_boot <- function(fit,
                         "Please fit the model with se = 'boot'"))
           }
       }
-    ft <- lavaan::lavInspect(boot_test, "timing")$total
+    ft <- tryCatch(lavaan::lavInspect(fit, "timing")$total,
+                   error = function(e) e)
+    if (inherits(ft, "error")) {
+        ft <- lavaan::lavInspect(boot_test, "timing")$total
+      }
     requireNamespace("parallel", quietly = TRUE)
     if (!is.null(seed)) set.seed(seed)
     if (ngp == 1) {
@@ -249,15 +253,21 @@ fit2boot_out_do_boot <- function(fit,
                                       make_cluster_args)},
                         error = function(e) e)
         has_cl <- !inherits(tmp, "error")
+        if (has_cl) {
+            on.exit(try(parallel::stopCluster(cl), silent = TRUE))
+          }
       } else {
         has_cl <- FALSE
       }
     if (has_cl) {
-        texp <-  1.2 * R * ft[[1]] / length(cl)
+        texp <-  2 * R * ft[[1]] / length(cl)
         message(paste0(length(cl), " processes started to run bootstrapping."))
-        message(paste0("The expected CPU time is about ",
-                        round(texp, 2),
-                        " second(s)."))
+        # No longer display the expected time if pbapply is used
+        if (!progress) {
+            message(paste0("The expected CPU time is at least ",
+                            round(texp, 2),
+                            " second(s) (can be much longer for some models)"))
+          }
         utils::flush.console()
         pkgs <- .packages()
         pkgs <- rev(pkgs)
@@ -286,7 +296,7 @@ fit2boot_out_do_boot <- function(fit,
             try(parallel::stopCluster(cl), silent = TRUE)
             stop("Running in parallel failed. Please set 'parallel' to FALSE.")
           }
-        parallel::stopCluster(cl)
+         try(parallel::stopCluster(cl), silent = TRUE)
       } else {
         if (progress) {
             rt <- system.time(out <- suppressWarnings(pbapply::pblapply(ids, boot_i,