+++ NEWS.md
@@ -1,4 +1,4 @@
-# manymome 0.2.4.5
+# manymome 0.2.4.9
 
 ## New Features
 
@@ -24,6 +24,12 @@
   consistent with the number of successful
   bootstrap replications. (0.2.4.2)
 
+- For functions such as `cond_effects()`,
+  if the model has only one outcome
+  variable (e.g., moderation only and
+  has no mediator), then the `y` argument
+  can be omitted. (0.2.4.6)
+
 ## Bugs
 
 - Remove a warning that may occur when
@@ -39,6 +45,19 @@
   `cond_indirect_effects`. (0.2.4.3)
 
 - Updated the vignettes (0.2.4.5).
+
+- Updated the internal `lavaan.mi` objects
+  used for tests. (0.2.4.7)
+
+- Add `on.exit()` to ensure that the clusters
+  created will be cleared if a function
+  is aborted (e.g., terminated by users).
+  (0.2.4.8)
+
+- Updated the way the expected CPU time
+  for bootstrapping, and the time is
+  displayed only if `progress` is
+  `FALSE`. (0.2.4.9)
 
 # manymome 0.2.4